refactor: Migrate tests from test_mutator.py to modular test structure
Dismantled the legacy monolithic test file (tests/test_mutator.py) and
migrated all 82 test methods to the corresponding modular test files in
tests/mutators/. This unifies the test suite architecture and improves
test organization.

Test distribution:
- test_generic.py: Added 18 tests for basic mutators (OperatorSwapper,
  ComparisonSwapper, ConstantPerturbator, GuardInjector, etc.)
- test_scenarios_control.py: Added 8 tests for control flow mutators
  (GuardExhaustionGenerator, ExitStresser, TraceBreaker, DeepCallMutator)
- test_scenarios_data.py: Added 11 tests for data structure mutators and
  helper functions (DictPolluter, MagicMethodMutator, attack generators)
- test_scenarios_runtime.py: Added 11 tests for runtime manipulation
  (StressPatternInjector, SideEffectInjector, GCInjector, etc.)
- test_scenarios_types.py: Added 10 tests for type system mutators
  (TypeInstabilityInjector, InlineCachePolluter, FunctionPatcher, etc.)
- test_utils.py: Added 18 tests for utility transformers and evil object
  generators (FuzzerSetupNormalizer, EmptyBodySanitizer, etc.)
- test_engine.py: Merged 10 tests for ASTMutator engine and integration
  scenarios into existing TestASTMutator class

All imports updated and validated. Test suite runs successfully with
342 tests passing (1 skipped) in tests/mutators/.

Fixes #194
