+++ tests/mutators/test_generic.py
@@ -18,14 +18,24 @@ from lafleur.mutators.generic import (
     AsyncConstructMutator,
     BlockTransposerMutator,
     BoundaryValuesMutator,
+    ComparisonSwapper,
+    ConstantPerturbator,
+    ContainerChanger,
     DecoratorMutator,
     ExceptionGroupMutator,
+    ForLoopInjector,
+    GuardInjector,
+    GuardRemover,
     NewUnpackingMutator,
+    OperatorSwapper,
     PatternMatchingMutator,
     SliceMutator,
+    StatementDuplicator,
     StringInterpolationMutator,
     SysMonitoringMutator,
     UnpackingMutator,
+    VariableRenamer,
+    VariableSwapper,
 )
 
 
@@ -1316,6 +1326,264 @@ class TestSysMonitoringMutator(unittest.TestCase):
         self.assertIn("def monitored_gym_7000", result)
         self.assertIn("for i in range(n):", result)
         self.assertIn("if i % 2 == 0:", result)
+
+
+class TestBasicMutators(unittest.TestCase):
+    """Test basic AST mutators."""
+
+    def setUp(self):
+        """Set random seed for reproducible tests."""
+        random.seed(42)
+
+    def test_operator_swapper(self):
+        """Test OperatorSwapper mutator."""
+        code = "result = a + b"
+        tree = ast.parse(code)
+
+        # Force mutation by patching random
+        with patch("random.random", return_value=0.1):
+            mutator = OperatorSwapper()
+            mutated = mutator.visit(tree)
+
+        # Check that the operator changed
+        binop = mutated.body[0].value
+        self.assertIsInstance(binop, ast.BinOp)
+        # Should not be Add anymore (with high probability)
+        # Note: Due to randomness, we can't guarantee which operator it becomes
+
+    def test_comparison_swapper(self):
+        """Test ComparisonSwapper mutator."""
+        code = "if a < b: pass"
+        tree = ast.parse(code)
+
+        with patch("random.random", return_value=0.3):
+            mutator = ComparisonSwapper()
+            mutated = mutator.visit(tree)
+
+        compare = mutated.body[0].test
+        self.assertIsInstance(compare, ast.Compare)
+        # Should have swapped the operator
+        self.assertIsInstance(compare.ops[0], ast.GtE)
+
+    def test_constant_perturbator_int(self):
+        """Test ConstantPerturbator on integers."""
+        code = "x = 42"
+        tree = ast.parse(code)
+
+        with patch("random.random", return_value=0.1):
+            with patch("random.choice", return_value=1):
+                mutator = ConstantPerturbator()
+                mutated = mutator.visit(tree)
+
+        constant = mutated.body[0].value
+        self.assertEqual(constant.value, 43)  # 42 + 1
+
+    def test_constant_perturbator_string(self):
+        """Test ConstantPerturbator on strings."""
+        code = "x = 'hello'"
+        tree = ast.parse(code)
+
+        with patch("random.random", return_value=0.1):
+            with patch("random.randint", return_value=1):  # Mutate 'e'
+                with patch("random.choice", return_value=1):
+                    mutator = ConstantPerturbator()
+                    mutated = mutator.visit(tree)
+
+        constant = mutated.body[0].value
+        self.assertEqual(len(constant.value), 5)
+        self.assertNotEqual(constant.value, "hello")
+
+    def test_guard_injector(self):
+        """Test GuardInjector mutator."""
+        code = "x = 1"
+        tree = ast.parse(code)
+
+        mutator = GuardInjector()
+        mutated = mutator.visit(tree)
+
+        # Should wrap in an If statement
+        self.assertIsInstance(mutated.body[0], ast.If)
+        # Check the test uses fuzzer_rng
+        test = mutated.body[0].test
+        self.assertIsInstance(test, ast.Compare)
+        self.assertIsInstance(test.left, ast.Call)
+        self.assertEqual(test.left.func.value.id, "fuzzer_rng")
+
+    def test_container_changer_list_to_set(self):
+        """Test ContainerChanger converting list to set."""
+        code = "x = [1, 2, 3]"
+        tree = ast.parse(code)
+
+        with patch("random.random", return_value=0.3):
+            mutator = ContainerChanger()
+            mutated = mutator.visit(tree)
+
+        container = mutated.body[0].value
+        self.assertIsInstance(container, ast.Set)
+
+    def test_container_changer_list_comprehension(self):
+        """Test ContainerChanger on list comprehension."""
+        code = "x = [i for i in range(10)]"
+        tree = ast.parse(code)
+
+        with patch("random.random", return_value=0.3):
+            mutator = ContainerChanger()
+            mutated = mutator.visit(tree)
+
+        comp = mutated.body[0].value
+        self.assertIsInstance(comp, ast.SetComp)
+
+    def test_variable_swapper(self):
+        """Test VariableSwapper mutator."""
+        code = dedent("""
+            x = 1
+            y = 2
+            z = x + y
+        """)
+        tree = ast.parse(code)
+
+        mutator = VariableSwapper()
+        mutated = mutator.visit(tree)
+
+        # Check that some swapping occurred
+        # Due to randomness, we can't predict exact swaps
+        self.assertIsInstance(mutated, ast.Module)
+
+    def test_variable_swapper_protected_names(self):
+        """Test VariableSwapper doesn't swap protected names."""
+        code = dedent("""
+            print(len([1, 2, 3]))
+            x = 1
+            y = 2
+        """)
+        tree = ast.parse(code)
+
+        mutator = VariableSwapper()
+        mutated = mutator.visit(tree)
+
+        # print and len should not be swapped
+        call = mutated.body[0].value
+        self.assertEqual(call.func.id, "print")
+        self.assertEqual(call.args[0].func.id, "len")
+
+    def test_statement_duplicator(self):
+        """Test StatementDuplicator mutator."""
+        code = "x = 1"
+        tree = ast.parse(code)
+
+        with patch("random.random", return_value=0.05):
+            mutator = StatementDuplicator()
+            mutated = mutator.visit(tree)
+
+        # Should have duplicated the statement
+        self.assertEqual(len(mutated.body), 2)
+        self.assertIsInstance(mutated.body[0], ast.Assign)
+        self.assertIsInstance(mutated.body[1], ast.Assign)
+
+    def test_variable_renamer(self):
+        """Test VariableRenamer mutator."""
+        code = "x = 1; y = x + 2"
+        tree = ast.parse(code)
+
+        remapping = {"x": "renamed_x", "y": "renamed_y"}
+        mutator = VariableRenamer(remapping)
+        mutated = mutator.visit(tree)
+
+        # Check variables were renamed
+        self.assertEqual(mutated.body[0].targets[0].id, "renamed_x")
+        self.assertEqual(mutated.body[1].targets[0].id, "renamed_y")
+        self.assertEqual(mutated.body[1].value.left.id, "renamed_x")
+
+    def test_for_loop_injector(self):
+        """Test ForLoopInjector mutator."""
+        code = dedent("""
+            x = 1
+            y = 2
+        """)
+        tree = ast.parse(code)
+
+        with patch("random.random", return_value=0.01):
+            with patch("random.choice", return_value="range"):
+                with patch("random.randint", return_value=100):
+                    mutator = ForLoopInjector()
+                    mutated = mutator.visit(tree)
+
+        # First statement should be wrapped in a loop
+        self.assertIsInstance(mutated.body[0], ast.For)
+
+    def test_guard_remover(self):
+        """Test GuardRemover mutator."""
+        code = dedent("""
+            if fuzzer_rng.random() < 0.1:
+                x = 1
+        """)
+        tree = ast.parse(code)
+
+        with patch("random.random", return_value=0.1):
+            mutator = GuardRemover()
+            mutated = mutator.visit(tree)
+
+        # Should have removed the guard
+        self.assertIsInstance(mutated.body[0], ast.Assign)
+        self.assertEqual(mutated.body[0].targets[0].id, "x")
+
+    def test_guard_injector_with_none_node(self):
+        """Test GuardInjector when child visitor returns None."""
+
+        # Create a custom node that will be removed
+        class RemovingTransformer(ast.NodeTransformer):
+            def visit_Assign(self, node):
+                return None  # Remove assignments
+
+        code = "x = 1"
+        tree = ast.parse(code)
+
+        # First apply removing transformer
+        remover = RemovingTransformer()
+        tree = remover.visit(tree)
+
+        # Then apply GuardInjector
+        injector = GuardInjector()
+        result = injector.visit(tree)
+
+        # Should handle None gracefully
+        self.assertIsInstance(result, ast.Module)
+
+    def test_variable_swapper_single_var(self):
+        """Test VariableSwapper with only one variable."""
+        code = "x = 1"
+        tree = ast.parse(code)
+
+        mutator = VariableSwapper()
+        mutated = mutator.visit(tree)
+
+        # Should handle gracefully without swapping
+        self.assertIsInstance(mutated, ast.Module)
+        self.assertEqual(mutated.body[0].targets[0].id, "x")
+
+    def test_for_loop_injector_with_del(self):
+        """Test ForLoopInjector doesn't wrap del statements."""
+        code = "del x"
+        tree = ast.parse(code)
+
+        with patch("random.random", return_value=0.01):
+            mutator = ForLoopInjector()
+            mutated = mutator.visit(tree)
+
+        # Should not wrap del in loop
+        self.assertIsInstance(mutated.body[0], ast.Delete)
+
+    def test_guard_injector_output(self):
+        """Test GuardInjector produces correct code."""
+        code = "x = 1"
+        tree = ast.parse(code)
+
+        mutator = GuardInjector()
+        mutated = mutator.visit(tree)
+
+        output = ast.unparse(mutated)
+        self.assertIn("if fuzzer_rng.random() < 0.1:", output)
+        self.assertIn("x = 1", output)
 
 
 if __name__ == "__main__":
